chatbot/aws-bedrock/text-chat/server/requirements.txt
# requirements.txt
fastapi
uvicorn
python-dotenv
boto3
botocore
